refactor: remove legacy panorama preview helpers
