+++ チェックポイント.md
@@ -1,52 +1,50 @@
 # チェックポイント
 
-このブランチでは、`coverage 180/360` の実装を先に閉じ、その後に panorama rendering の shared pipeline refactor を段階的に進める。
+この仕上げでは、consumer の移行後に残っている legacy render 依存を shared render 層から順に除去する。
 
 ## 現在地点
 
-- `coverage` は backend / modal / preview に入っている
-- ただし rendering architecture は未統一
-- `stickers` だけ fast path、`cutout` と `preview` は旧経路が残る
+- [x] `modal pano`
+- [x] `cutout preview`
+- [x] `preview runtime`
+- [x] `preview node`
 
-## 実装フェーズ
+上の 4 consumer は新 core 経由で描画している。
 
-### Phase 0: coverage branch を閉じる
+残っているのは、shared 層にぶら下がった旧 helper の整理だけ。
 
-- [x] `coverage` を node parameter として追加
-- [x] modal inspector と node widget を連携
-- [x] background / preview / output に `coverage` を反映
-- [x] 設計メモを `docs/adr/0014-unified-panorama-render-pipeline.md` に記録
-- [ ] 現在の working tree を checkpoint commit
+## 残タスク
 
-### Phase 1: 共有 pipeline の土台を整理
+### 1. preview helper の整理
 
-- [ ] panorama background / object / overlay の 3 pass 契約を明文化
-- [ ] view builder / scene builder / texture builder の責務を集約
-- [ ] `renderCutoutViewToContext2D(...)` 依存箇所を洗い出し、 shared pipeline へ寄せる準備をする
+- [x] `pano_preview_render.js` から純関数を退避
+- [x] `pano_preview_previewnode.js` の `pano_preview_render.js` 依存を除去
+- [x] `pano_preview_render.js` を削除
 
-### Phase 2: cutout object pass を shared pipeline へ移す
+### 2. background pass の legacy 依存除去
 
-- [ ] `cutout` pano view で frame/object が重くなる経路を shared object pass へ移す
-- [ ] `stickers` 専用 fast path を generic pano layer renderer に寄せる
-- [ ] `cutout` 背景 + object の render policy を `stickers` と揃える
+- [x] `pano_render_background_pass.js` から `pano_gl_viewport.js` import を除去
+- [x] background pass は依存注入がある時だけ外部 renderer を使う形に整理
 
-### Phase 3: preview orchestration を shared pipeline consumer に揃える
+### 3. legacy viewport helper の削除
 
-- [ ] node preview
-- [ ] standalone preview node
-- [ ] read-only modal preview
-- [ ] それぞれが shared pipeline の consumer になるよう整理
+- [x] `pano_gl_viewport.js` の実参照を 0 にする
+- [x] `pano_gl_viewport.js` を削除
 
-### Phase 4: 残件整理
+### 4. テスト整理
 
-- [ ] profiling / debug hook を整理
-- [ ] 不要 helper / 重複 render path / 一時 adapter を削除
-- [ ] 生成物 build
-- [ ] 最終 cleanup と commit
+- [x] 旧 helper 存在前提のテストを新構造へ更新
+- [x] core / consumer / cleanup 後の最低限テストを通す
+
+### 5. 最終確認
+
+- [x] consumer から legacy helper 名が 0 件
+- [x] shared 層から `pano_gl_viewport.js` 依存が 0 件
+- [ ] `coverage` が background のみに作用
+- [ ] build 反映
 
 ## 原則
 
-- `stickers` / `cutout` / `preview` は同じ panorama pipeline に乗せる
-- `frame` / `unwrap` は別責務として無理に混ぜない
-- `video` / `3D` は「背景 source」か「配置 object」かで pass を分ける
-- node id / port semantics は変えない
+- fallback は残さない
+- 未実装機能は無理に混ぜない
+- 旧 helper を温存して「動いて見える」状態にはしない
